Added a GIF of the game to my README.md.
I had trouble getting my video to play; I uploaded it onto imgur and used the URL like some other examples did, but it only displayed as a photo on GitHub. I couldn't get the .mp4 file to display in Markdown at all, so I used an online video converter to turn it into a .gif file, and it now plays the full clip.